backend/analytics-engine/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
kafka-python==2.0.2
boto3==1.26.0
azure-ai-ml==1.8.0
scikit-learn==1.2.2
pandas==2.0.0
numpy==1.24.2